+++ ExampleMod/Common/Configs/ModConfigShowcases/ModConfigShowcaseAccessibility.cs
@@ -9,7 +9,7 @@ using ExampleMod.Common.Configs.CustomDataTypes;
 
 // Because this config was designed to show off various UI capabilities,
 // this config have no effect on the mod and provides purely teaching example.
-namespace ExampleMod.Common.Configs
+namespace ExampleMod.Common.Configs.ModConfigShowcases
 {
 	[BackgroundColor(164, 153, 190)]
 	[Label("ModConfig Showcase F: Accessibility")]
