+++ ExampleMod/Common/Configs/ModConfigShowcases/ModConfigShowcaseDefaultValues.cs
@@ -10,7 +10,7 @@ using ExampleMod.Common.Configs.CustomDataTypes;
 
 // Because this config was designed to show off various UI capabilities,
 // this config have no effect on the mod and provides purely teaching example.
-namespace ExampleMod.Common.Configs
+namespace ExampleMod.Common.Configs.ModConfigShowcases
 {
 	[BackgroundColor(164, 153, 190)]
 	[Label("ModConfig Showcase D: Default Values")]
