+++ ExampleMod/Common/Configs/ModConfigShowcases/ModConfigShowcaseLabels.cs
@@ -7,7 +7,7 @@ using ExampleMod.Common.Configs.CustomDataTypes;
 
 // Because this config was designed to show off various UI capabilities,
 // this config have no effect on the mod and provides purely teaching example.
-namespace ExampleMod.Common.Configs
+namespace ExampleMod.Common.Configs.ModConfigShowcases
 {
 	[BackgroundColor(154, 152, 181)]
 	[Label("ModConfig Showcase C: Labels")]
